Reorganize classes
- Move base classes to its own package
- Move most logic out of MagiskActivity to MainActivity
