+++ app/src/main/java/com/topjohnwu/magisk/base/MagiskFragment.kt
@@ -1,0 +1,51 @@
+package com.topjohnwu.magisk.base
+
+import android.os.Bundle
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+import androidx.annotation.CallSuper
+import androidx.databinding.DataBindingUtil
+import androidx.databinding.ViewDataBinding
+import androidx.fragment.app.Fragment
+import com.topjohnwu.magisk.BR
+import com.topjohnwu.magisk.base.viewmodel.MagiskViewModel
+import com.topjohnwu.magisk.model.events.EventHandler
+import com.topjohnwu.magisk.model.events.ViewEvent
+import com.topjohnwu.magisk.ui.MainActivity
+
+abstract class MagiskFragment<ViewModel : MagiskViewModel, Binding : ViewDataBinding> :
+    Fragment(), EventHandler {
+
+    protected val activity get() = requireActivity() as MainActivity
+    protected lateinit var binding: Binding
+    protected abstract val layoutRes: Int
+    protected abstract val viewModel: ViewModel
+
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        viewModel.viewEvents.observe(this, viewEventObserver)
+    }
+
+    override fun onCreateView(
+            inflater: LayoutInflater,
+            container: ViewGroup?,
+            savedInstanceState: Bundle?
+    ): View? {
+        binding = DataBindingUtil.inflate<Binding>(inflater, layoutRes, container, false).apply {
+            setVariable(BR.viewModel, viewModel)
+            lifecycleOwner = this@MagiskFragment
+        }
+
+        return binding.root
+    }
+
+    @CallSuper
+    override fun onEventDispatched(event: ViewEvent) {
+        super.onEventDispatched(event)
+        activity.onEventDispatched(event)
+    }
+
+    open fun onBackPressed(): Boolean = false
+
+}
