+++ app/src/main/java/com/topjohnwu/magisk/model/events/SimpleViewEvent.kt
@@ -1,7 +1,5 @@
 package com.topjohnwu.magisk.model.events
 
-import com.topjohnwu.magisk.model.events.ViewEvent
-
 class SimpleViewEvent(
     val event: Int
 ) : ViewEvent()
