+++ app/src/main/java/com/topjohnwu/magisk/ui/log/LogFragment.kt
@@ -6,11 +6,11 @@ import android.view.Menu
 import android.view.MenuInflater
 import android.view.MenuItem
 import android.view.View
-import com.topjohnwu.magisk.model.events.ViewEvent
 import com.topjohnwu.magisk.R
+import com.topjohnwu.magisk.base.MagiskFragment
 import com.topjohnwu.magisk.databinding.FragmentLogBinding
 import com.topjohnwu.magisk.model.events.PageChangedEvent
-import com.topjohnwu.magisk.ui.base.MagiskFragment
+import com.topjohnwu.magisk.model.events.ViewEvent
 import org.koin.androidx.viewmodel.ext.android.viewModel
 
 class LogFragment : MagiskFragment<LogViewModel, FragmentLogBinding>() {
