+++ app/src/main/java/com/topjohnwu/magisk/ui/superuser/SuperuserFragment.kt
@@ -1,8 +1,8 @@
 package com.topjohnwu.magisk.ui.superuser
 
 import com.topjohnwu.magisk.R
+import com.topjohnwu.magisk.base.MagiskFragment
 import com.topjohnwu.magisk.databinding.FragmentSuperuserBinding
-import com.topjohnwu.magisk.ui.base.MagiskFragment
 import org.koin.androidx.viewmodel.ext.android.viewModel
 
 class SuperuserFragment :
