+++ app/src/main/java/com/topjohnwu/magisk/ui/surequest/SuRequestActivity.kt
@@ -5,13 +5,13 @@ import android.os.Build
 import android.os.Bundle
 import android.text.TextUtils
 import android.view.Window
-import com.topjohnwu.magisk.model.events.ViewEvent
 import com.topjohnwu.magisk.R
+import com.topjohnwu.magisk.base.MagiskActivity
 import com.topjohnwu.magisk.databinding.ActivityRequestBinding
 import com.topjohnwu.magisk.model.entity.MagiskPolicy
 import com.topjohnwu.magisk.model.events.DieEvent
+import com.topjohnwu.magisk.model.events.ViewEvent
 import com.topjohnwu.magisk.model.receiver.GeneralReceiver
-import com.topjohnwu.magisk.ui.base.MagiskActivity
 import com.topjohnwu.magisk.utils.SuLogger
 import org.koin.androidx.viewmodel.ext.android.viewModel
